# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IoT Vision Hub",
    "summary": "Asynchronous camera monitoring: motion events, alerts and live streams.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/push": {
      "get": {
        "tags": [
          "push"
        ],
        "summary": "Get the push key",
        "description": "The hub's public VAPID key, to pass to `pushManager.subscribe()` as the\n`applicationServerKey`. 404 when push is turned off on the hub.",
        "operationId": "push_get_push_config",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushConfigOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/push/subscriptions": {
      "post": {
        "tags": [
          "push"
        ],
        "summary": "Subscribe this browser",
        "description": "Send `PushSubscription.toJSON()`. Subscribing the same browser again replaces its\nentry, under the account signed in now. Endpoints must be on a known push service.",
        "operationId": "push_subscribe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscriptionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushSubscriptionOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/push/subscriptions/{subscription_id}": {
      "delete": {
        "tags": [
          "push"
        ],
        "summary": "Unsubscribe a browser",
        "description": "The id is the SHA-256 of the endpoint (hex). Only your own subscriptions.",
        "operationId": "push_unsubscribe",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{64}$",
              "title": "Subscription Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/push/test": {
      "post": {
        "tags": [
          "push"
        ],
        "summary": "Send a test notification",
        "description": "To every browser subscribed under your account.",
        "operationId": "push_send_test",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushTestOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "description": "One request validation failure. The offending input is deliberately omitted.",
        "properties": {
          "loc": {
            "description": "Path to the invalid field, e.g. ['body', 'port']",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "ProblemDetail": {
        "additionalProperties": true,
        "properties": {
          "type": {
            "default": "about:blank",
            "description": "URI identifying the problem type",
            "examples": [
              "about:blank"
            ],
            "title": "Type",
            "type": "string"
          },
          "title": {
            "description": "Short, human-readable summary of the problem type",
            "title": "Title",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "maximum": 599,
            "minimum": 400,
            "title": "Status",
            "type": "integer"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Explanation specific to this occurrence",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Request path that produced the problem",
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Correlates with server logs",
            "title": "Request Id"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Validation failures",
            "title": "Errors"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
      "PushConfigOut": {
        "properties": {
          "public_key": {
            "type": "string",
            "title": "Public Key",
            "description": "`applicationServerKey` to subscribe with"
          }
        },
        "type": "object",
        "required": [
          "public_key"
        ],
        "title": "PushConfigOut"
      },
      "PushKeysIn": {
        "properties": {
          "p256dh": {
            "type": "string",
            "maxLength": 100,
            "pattern": "^[A-Za-z0-9_-]+={0,2}$",
            "title": "P256Dh"
          },
          "auth": {
            "type": "string",
            "maxLength": 50,
            "pattern": "^[A-Za-z0-9_-]+={0,2}$",
            "title": "Auth"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "p256dh",
          "auth"
        ],
        "title": "PushKeysIn"
      },
      "PushSubscriptionIn": {
        "properties": {
          "endpoint": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 12,
            "title": "Endpoint"
          },
          "keys": {
            "$ref": "#/components/schemas/PushKeysIn"
          },
          "expirationTime": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expirationtime"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "endpoint",
          "keys"
        ],
        "title": "PushSubscriptionIn"
      },
      "PushSubscriptionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "SHA-256 of the endpoint, hex"
          },
          "created_at": {
            "$ref": "#/components/schemas/UtcDateTime"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at"
        ],
        "title": "PushSubscriptionOut"
      },
      "PushTestOut": {
        "properties": {
          "delivered": {
            "type": "integer",
            "title": "Delivered",
            "description": "Browsers whose push service accepted it"
          }
        },
        "type": "object",
        "required": [
          "delivered"
        ],
        "title": "PushTestOut"
      },
...
      "UtcDateTime": {
        "type": "string",
        "format": "date-time"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        }
      }
    }
  },
...
}